Added getters for tester Valente (references #9 #14)

+++ tax/src/main/java/pt/ulisboa/tecnico/softeng/tax/domain/ItemType.java
@@ -38,7 +38,7 @@ public class ItemType {
 		return this._name;
 	}
 
-	public float getTax() {
+	public int getTax() {
 		return this._tax;
 	}
